+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 == Changelog ==
 
+= 1.0.8 =
+* Fixed author filter to use theme Filter
+* Added author filter to Proceedings archive 
+
 = 1.0.7 =
 * Single Proceeding Pagination now sorted by session date and time
 * Author now only shows for non-users (bylines)